+++ src/mock/event-mock.js
@@ -1,17 +1,11 @@
 import dayjs from 'dayjs';
-import minMax from 'dayjs/plugin/minMax.js';
+import minMax from 'dayjs/plugin/minMax';
+import {getRandomInteger} from '../utils/common';
 dayjs.extend(minMax);
 
 const MAX_MINUTE_GAP = 30;
 const MAX_HOUR_GAP = 3;
 const MAX_DAY_GAP = 3;
-
-const getRandomInteger = (a = 0, b = 1) => {
-  const lower = Math.ceil(Math.min(a, b));
-  const upper = Math.floor(Math.max(a, b));
-
-  return Math.floor(lower + Math.random() * (upper - lower + 1));
-};
 
 const descriptions = ['Lorem ipsum dolor sit amet, consectetur adipiscing elit.',
   'Cras aliquet varius magna, non porta ligula feugiat eget.',
