+++ src/view/events-list.js
@@ -1,4 +1,4 @@
-import {createElement} from '../utils';
+import AbstractView from './abstract';
 
 const createEventsListTemplate = () => (
   `<ul class="trip-events__list">
@@ -6,25 +6,9 @@ const createEventsListTemplate = () => (
 </ul>`
 );
 
-class EventsList {
-  constructor() {
-    this._element = null;
-  }
-
+class EventsList extends AbstractView {
   getTemplate() {
     return createEventsListTemplate();
-  }
-
-  getElement() {
-    if (!this._element) {
-      this._element = createElement(this.getTemplate());
-    }
-
-    return this._element;
-  }
-
-  removeElement() {
-    this._element = null;
   }
 }
