+++ src/view/no-events.js
@@ -1,4 +1,4 @@
-import {createElement} from '../utils';
+import AbstractView from './abstract';
 
 const createNoEventsTemplate = () => (
   `<p class="trip-events__msg">
@@ -6,25 +6,9 @@ const createNoEventsTemplate = () => (
   </p>`
 );
 
-class NoEvent {
-  constructor() {
-    this._element = null;
-  }
-
+class NoEvent extends AbstractView {
   getTemplate() {
     return createNoEventsTemplate();
-  }
-
-  getElement() {
-    if (!this._element) {
-      this._element = createElement(this.getTemplate());
-    }
-
-    return this._element;
-  }
-
-  removeElement() {
-    this._element = null;
   }
 }
